rewrote the logitech device
