+++ Project-Aurora/Project-Aurora/Devices/Logitech/LogitechGSDK.cs
@@ -1,13 +1,10 @@
-﻿using System;
-using System.Collections;
+﻿
+using LedCSharp;
+using System;
+using System.Drawing;
 using System.Runtime.InteropServices;
 using System.Text;
-using System.IO;
-using System.Reflection;
-using Aurora;
-using System.ComponentModel;
 
-#pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
 namespace LedCSharp
 {
     public enum keyboardNames
@@ -27,7 +24,7 @@ namespace LedCSharp
         F12 = 0x58,
         PRINT_SCREEN = 0x137,
         SCROLL_LOCK = 0x46,
-        PAUSE_BREAK = 0x45,
+        PAUSE_BREAK = 0x145,
         TILDE = 0x29,
         ONE = 0x02,
         TWO = 0x03,
@@ -45,7 +42,7 @@ namespace LedCSharp
         INSERT = 0x152,
         HOME = 0x147,
         PAGE_UP = 0x149,
-        NUM_LOCK = 0x145,
+        NUM_LOCK = 0x45,
         NUM_SLASH = 0x135,
         NUM_ASTERISK = 0x37,
         NUM_MINUS = 0x4A,
@@ -129,49 +126,159 @@ namespace LedCSharp
         G_BADGE = 0xFFFF2
     };
 
-    public class LogitechGSDK
+    public enum Logitech_keyboardBitmapKeys
     {
-        public enum LGDLL
-        {
-            [Description("LGS")]
-            LGS,
-            [Description("GHUB")]
-            GHUB
-        }
+        UNKNOWN = -1,
+        ESC = 0,
+        F1 = 4,
+        F2 = 8,
+        F3 = 12,
+        F4 = 16,
+        F5 = 20,
+        F6 = 24,
+        F7 = 28,
+        F8 = 32,
+        F9 = 36,
+        F10 = 40,
+        F11 = 44,
+        F12 = 48,
+        PRINT_SCREEN = 52,
+        SCROLL_LOCK = 56,
+        PAUSE_BREAK = 60,
+        //64
+        //68
+        //72
+        //76
+        //80
 
-        [DllImport("kernel32.dll", CharSet = CharSet.Auto, SetLastError = true)]
-        private static extern bool SetDllDirectory(string path);
+        TILDE = 84,
+        ONE = 88,
+        TWO = 92,
+        THREE = 96,
+        FOUR = 100,
+        FIVE = 104,
+        SIX = 108,
+        SEVEN = 112,
+        EIGHT = 116,
+        NINE = 120,
+        ZERO = 124,
+        MINUS = 128,
+        EQUALS = 132,
+        BACKSPACE = 136,
+        INSERT = 140,
+        HOME = 144,
+        PAGE_UP = 148,
+        NUM_LOCK = 152,
+        NUM_SLASH = 156,
+        NUM_ASTERISK = 160,
+        NUM_MINUS = 164,
 
-        static LogitechGSDK()
-        {
-            //InitDLL(LGDLL.GHUB);
-        }
+        TAB = 168,
+        Q = 172,
+        W = 176,
+        E = 180,
+        R = 184,
+        T = 188,
+        Y = 192,
+        U = 196,
+        I = 200,
+        O = 204,
+        P = 208,
+        OPEN_BRACKET = 212,
+        CLOSE_BRACKET = 216,
+        BACKSLASH = 220,
+        KEYBOARD_DELETE = 224,
+        END = 228,
+        PAGE_DOWN = 232,
+        NUM_SEVEN = 236,
+        NUM_EIGHT = 240,
+        NUM_NINE = 244,
+        NUM_PLUS = 248,
 
-        private static bool initd = false;
-        public static void InitDLL(LGDLL? dll = null)
-        {
-            if (initd)
-                return;
+        CAPS_LOCK = 252,
+        A = 256,
+        S = 260,
+        D = 264,
+        F = 268,
+        G = 272,
+        H = 276,
+        J = 280,
+        K = 284,
+        L = 288,
+        SEMICOLON = 292,
+        APOSTROPHE = 296,
+        HASHTAG = 300,//300
+        ENTER = 304,
+        //308
+        //312
+        //316
+        NUM_FOUR = 320,
+        NUM_FIVE = 324,
+        NUM_SIX = 328,
+        //332
 
-            bool ghub = false;
-            var path = Path.GetDirectoryName(Assembly.GetEntryAssembly().Location);
+        LEFT_SHIFT = 336,
+        BACKSLASH_UK = 340,
+        Z = 344,
+        X = 348,
+        C = 352,
+        V = 356,
+        B = 360,
+        N = 364,
+        M = 368,
+        COMMA = 372,
+        PERIOD = 376,
+        FORWARD_SLASH = 380,
+        OEM102 = 384,
+        RIGHT_SHIFT = 388,
+        //392
+        ARROW_UP = 396,
+        //400
+        NUM_ONE = 404,
+        NUM_TWO = 408,
+        NUM_THREE = 412,
+        NUM_ENTER = 416,
 
-            if (dll == null)
-                ghub = Directory.Exists(Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ProgramFiles), "LGHUB"));
-            else
-                ghub = dll.Equals(LGDLL.GHUB);
+        LEFT_CONTROL = 420,
+        LEFT_WINDOWS = 424,
+        LEFT_ALT = 428,
+        //432
+        JPN_MUHENKAN = 436,//436
+        SPACE = 440,
+        //444
+        //448
+        JPN_HENKAN = 452,//452
+        JPN_HIRAGANA_KATAKANA = 456,//456
+        //460
+        RIGHT_ALT = 464,
+        RIGHT_WINDOWS = 468,
+        APPLICATION_SELECT = 472,
+        RIGHT_CONTROL = 476,
+        ARROW_LEFT = 480,
+        ARROW_DOWN = 484,
+        ARROW_RIGHT = 488,
+        NUM_ZERO = 492,
+        NUM_PERIOD = 496,
+        //500
+    };
 
-            if (ghub)
-                Global.logger.LogLine("Loading GHUB DLL", Logging_Level.Info);
-            else
-                Global.logger.LogLine("Loading LGS DLL", Logging_Level.Info);
-            path = Path.Combine(path, "Logi", ghub ? "GHUB" : "LGS");
-            bool ok = SetDllDirectory(path);
-            if (!ok) throw new System.ComponentModel.Win32Exception();
-            initd = true;
-        }
+    public enum DeviceType
+    {
+        Keyboard = 0x0,
+        Mouse = 0x3,
+        Mousemat = 0x4,
+        Headset = 0x8,
+        Speaker = 0xe
+    }
 
-        //LED SDK
+    public enum LGDLL
+    {
+        LGS,
+        GHUB
+    }
+
+    public static class LogitechGSDK
+    {
         private const int LOGI_DEVICETYPE_MONOCHROME_ORD = 0;
         private const int LOGI_DEVICETYPE_RGB_ORD = 1;
         private const int LOGI_DEVICETYPE_PERKEY_RGB_ORD = 2;
@@ -188,87 +295,264 @@ namespace LedCSharp
         public const int LOGI_LED_BITMAP_SIZE = LOGI_LED_BITMAP_WIDTH * LOGI_LED_BITMAP_HEIGHT * LOGI_LED_BITMAP_BYTES_PER_KEY;
         public const int LOGI_LED_DURATION_INFINITE = 0;
 
-        [DllImport("LogitechLed ", CallingConvention = CallingConvention.Cdecl)]
+        public static bool GHUB = true;
+
+        public static bool LogiLedInit()
+        {
+            return GHUB ? GHUBImports.LogiLedInit() : LGSImports.LogiLedInit();
+        }
+
+        public static bool LogiLedSetTargetDevice(int targetDevice)
+        {
+            return GHUB ? GHUBImports.LogiLedSetTargetDevice(targetDevice) : LGSImports.LogiLedSetTargetDevice(targetDevice);
+        }
+
+        public static bool LogiLedSetLighting(Color color)
+        {
+            var (R, G, B) = GetColorValues(color);
+            return GHUB ? GHUBImports.LogiLedSetLighting(R, G, B) : LGSImports.LogiLedSetLighting(R, G, B);
+        }
+
+        public static bool LogiLedExcludeKeysFromBitmap(keyboardNames[] keyList, int listCount)
+        {
+            return GHUB ? GHUBImports.LogiLedExcludeKeysFromBitmap(keyList, listCount) :
+                     LGSImports.LogiLedExcludeKeysFromBitmap(keyList, listCount);
+        }
+
+        public static bool LogiLedSetLightingFromBitmap(byte[] bitmap)
+        {
+            return GHUB ? GHUBImports.LogiLedSetLightingFromBitmap(bitmap) :
+                     LGSImports.LogiLedSetLightingFromBitmap(bitmap);
+        }
+
+        public static bool LogiLedSetLightingForKeyWithScanCode(int keyCode, Color color)
+        {
+            var (R, G, B) = GetColorValues(color);
+            return GHUB ? GHUBImports.LogiLedSetLightingForKeyWithScanCode(keyCode, R, G, B) :
+                LGSImports.LogiLedSetLightingForKeyWithScanCode(keyCode, R, G, B);
+        }
+
+        public static bool LogiLedSetLightingForKeyWithHidCode(int keyCode, Color color)
+        {
+            var (R, G, B) = GetColorValues(color);
+            return GHUB ? GHUBImports.LogiLedSetLightingForKeyWithHidCode(keyCode, R, G, B) :
+                LGSImports.LogiLedSetLightingForKeyWithHidCode(keyCode, R, G, B);
+        }
+
+        public static bool LogiLedSetLightingForKeyWithQuartzCode(int keyCode, Color color)
+        {
+            var (R, G, B) = GetColorValues(color);
+            return GHUB ? GHUBImports.LogiLedSetLightingForKeyWithQuartzCode(keyCode, R, G, B) :
+                LGSImports.LogiLedSetLightingForKeyWithQuartzCode(keyCode, R, G, B);
+        }
+
+        public static bool LogiLedSetLightingForKeyWithKeyName(keyboardNames keyCode, Color color)
+        {
+            var (R, G, B) = GetColorValues(color);
+            return GHUB ? GHUBImports.LogiLedSetLightingForKeyWithKeyName(keyCode, R, G, B) :
+                LGSImports.LogiLedSetLightingForKeyWithKeyName(keyCode, R, G, B);
+        }
+
+        public static bool LogiLedSetLightingForTargetZone(DeviceType deviceType, int zone, Color color)
+        {
+            var (R, G, B) = GetColorValues(color);
+            return GHUB ? GHUBImports.LogiLedSetLightingForTargetZone((byte)deviceType, zone, R, G, B) :
+                LGSImports.LogiLedSetLightingForTargetZone((byte)deviceType, zone, R, G, B);
+        }
+
+        public static void LogiLedShutdown()
+        {
+            if (GHUB)
+                GHUBImports.LogiLedShutdown();
+            else
+                LGSImports.LogiLedShutdown();
+        }
+
+        public static bool LogiLedSaveCurrentLighting()
+        {
+            return GHUB ? GHUBImports.LogiLedSaveCurrentLighting() : LGSImports.LogiLedSaveCurrentLighting();
+        }
+
+        public static bool LogiLedRestoreLighting()
+        {
+            return GHUB ? GHUBImports.LogiLedRestoreLighting() : LGSImports.LogiLedRestoreLighting();
+        }
+
+        private static (int R, int G, int B) GetColorValues(Color clr)
+        {
+            return ((int)(clr.R / 255.0 * 100.0),
+                    (int)(clr.G / 255.0 * 100.0),
+                    (int)(clr.B / 255.0 * 100.0));
+        }
+    }
+
+    internal static class LGSImports
+    {
+        private const string dllpath = "Logi\\LGS\\LogitechLed.dll";
+
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
         public static extern bool LogiLedInit();
 
         //Config option functions
-        [DllImport("LogitechLed ", CallingConvention = CallingConvention.Cdecl)]
-        public static extern bool LogiLedGetConfigOptionNumber([MarshalAs(UnmanagedType.LPWStr)]String configPath, ref double defaultNumber);
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
+        public static extern bool LogiLedGetConfigOptionNumber([MarshalAs(UnmanagedType.LPWStr)] String configPath, ref double defaultNumber);
 
-        [DllImport("LogitechLed ", CallingConvention = CallingConvention.Cdecl)]
-        public static extern bool LogiLedGetConfigOptionBool([MarshalAs(UnmanagedType.LPWStr)]String configPath, ref bool defaultRed);
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
+        public static extern bool LogiLedGetConfigOptionBool([MarshalAs(UnmanagedType.LPWStr)] String configPath, ref bool defaultRed);
 
-        [DllImport("LogitechLed ", CallingConvention = CallingConvention.Cdecl)]
-        public static extern bool LogiLedGetConfigOptionColor([MarshalAs(UnmanagedType.LPWStr)]String configPath, ref int defaultRed, ref int defaultGreen, ref int defaultBlue);
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
+        public static extern bool LogiLedGetConfigOptionColor([MarshalAs(UnmanagedType.LPWStr)] String configPath, ref int defaultRed, ref int defaultGreen, ref int defaultBlue);
 
-        [DllImport("LogitechLed ", CallingConvention = CallingConvention.Cdecl)]
-        public static extern bool LogiLedGetConfigOptionKeyInput([MarshalAs(UnmanagedType.LPWStr)]String configPath, StringBuilder buffer, int bufsize);
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
+        public static extern bool LogiLedGetConfigOptionKeyInput([MarshalAs(UnmanagedType.LPWStr)] String configPath, StringBuilder buffer, int bufsize);
         /////////////////////
 
-        [DllImport("LogitechLed ", CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
         public static extern bool LogiLedSetTargetDevice(int targetDevice);
 
-        [DllImport("LogitechLed ", CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
         public static extern bool LogiLedGetSdkVersion(ref int majorNum, ref int minorNum, ref int buildNum);
 
-        [DllImport("LogitechLed ", CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
         public static extern bool LogiLedSaveCurrentLighting();
 
-        [DllImport("LogitechLed ", CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
         public static extern bool LogiLedSetLighting(int redPercentage, int greenPercentage, int bluePercentage);
 
-        [DllImport("LogitechLed ", CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
         public static extern bool LogiLedRestoreLighting();
 
-        [DllImport("LogitechLed ", CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
         public static extern bool LogiLedFlashLighting(int redPercentage, int greenPercentage, int bluePercentage, int milliSecondsDuration, int milliSecondsInterval);
 
-        [DllImport("LogitechLed ", CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
         public static extern bool LogiLedPulseLighting(int redPercentage, int greenPercentage, int bluePercentage, int milliSecondsDuration, int milliSecondsInterval);
 
-        [DllImport("LogitechLed ", CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
         public static extern bool LogiLedStopEffects();
 
-        [DllImport("LogitechLed ", CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
         public static extern bool LogiLedExcludeKeysFromBitmap(keyboardNames[] keyList, int listCount);
 
-        [DllImport("LogitechLed ", CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
         public static extern bool LogiLedSetLightingFromBitmap(byte[] bitmap);
 
-        [DllImport("LogitechLed ", CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
         public static extern bool LogiLedSetLightingForKeyWithScanCode(int keyCode, int redPercentage, int greenPercentage, int bluePercentage);
 
-        [DllImport("LogitechLed ", CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
         public static extern bool LogiLedSetLightingForKeyWithHidCode(int keyCode, int redPercentage, int greenPercentage, int bluePercentage);
 
-        [DllImport("LogitechLed ", CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
         public static extern bool LogiLedSetLightingForKeyWithQuartzCode(int keyCode, int redPercentage, int greenPercentage, int bluePercentage);
 
-        [DllImport("LogitechLed ", CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
         public static extern bool LogiLedSetLightingForKeyWithKeyName(keyboardNames keyCode, int redPercentage, int greenPercentage, int bluePercentage);
 
-        [DllImport("LogitechLed ", CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
         public static extern bool LogiLedSetLightingForTargetZone(byte deviceType, int zone, int redPercentage, int greenPercentage, int bluePercentage);
 
-        [DllImport("LogitechLed ", CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
         public static extern bool LogiLedSaveLightingForKey(keyboardNames keyName);
 
-        [DllImport("LogitechLed ", CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
         public static extern bool LogiLedRestoreLightingForKey(keyboardNames keyName);
 
-        [DllImport("LogitechLed ", CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
         public static extern bool LogiLedFlashSingleKey(keyboardNames keyName, int redPercentage, int greenPercentage, int bluePercentage, int msDuration, int msInterval);
 
-        [DllImport("LogitechLed ", CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
         public static extern bool LogiLedPulseSingleKey(keyboardNames keyName, int startRedPercentage, int startGreenPercentage, int startBluePercentage, int finishRedPercentage, int finishGreenPercentage, int finishBluePercentage, int msDuration, bool isInfinite);
 
-        [DllImport("LogitechLed ", CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
         public static extern bool LogiLedStopEffectsOnKey(keyboardNames keyName);
 
-        [DllImport("LogitechLed ", CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
         public static extern void LogiLedShutdown();
     }
 
-}
+    internal static class GHUBImports
+    {
+        private const string dllpath = "Logi\\GHUB\\LogitechLed.dll";
 
-#pragma warning restore CS1591 // Missing XML comment for publicly visible type or member
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
+        public static extern bool LogiLedInit();
+
+        //Config option functions
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
+        public static extern bool LogiLedGetConfigOptionNumber([MarshalAs(UnmanagedType.LPWStr)] String configPath, ref double defaultNumber);
+
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
+        public static extern bool LogiLedGetConfigOptionBool([MarshalAs(UnmanagedType.LPWStr)] String configPath, ref bool defaultRed);
+
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
+        public static extern bool LogiLedGetConfigOptionColor([MarshalAs(UnmanagedType.LPWStr)] String configPath, ref int defaultRed, ref int defaultGreen, ref int defaultBlue);
+
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
+        public static extern bool LogiLedGetConfigOptionKeyInput([MarshalAs(UnmanagedType.LPWStr)] String configPath, StringBuilder buffer, int bufsize);
+        /////////////////////
+
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
+        public static extern bool LogiLedSetTargetDevice(int targetDevice);
+
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
+        public static extern bool LogiLedGetSdkVersion(ref int majorNum, ref int minorNum, ref int buildNum);
+
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
+        public static extern bool LogiLedSaveCurrentLighting();
+
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
+        public static extern bool LogiLedSetLighting(int redPercentage, int greenPercentage, int bluePercentage);
+
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
+        public static extern bool LogiLedRestoreLighting();
+
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
+        public static extern bool LogiLedFlashLighting(int redPercentage, int greenPercentage, int bluePercentage, int milliSecondsDuration, int milliSecondsInterval);
+
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
+        public static extern bool LogiLedPulseLighting(int redPercentage, int greenPercentage, int bluePercentage, int milliSecondsDuration, int milliSecondsInterval);
+
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
+        public static extern bool LogiLedStopEffects();
+
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
+        public static extern bool LogiLedExcludeKeysFromBitmap(LedCSharp.keyboardNames[] keyList, int listCount);
+
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
+        public static extern bool LogiLedSetLightingFromBitmap(byte[] bitmap);
+
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
+        public static extern bool LogiLedSetLightingForKeyWithScanCode(int keyCode, int redPercentage, int greenPercentage, int bluePercentage);
+
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
+        public static extern bool LogiLedSetLightingForKeyWithHidCode(int keyCode, int redPercentage, int greenPercentage, int bluePercentage);
+
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
+        public static extern bool LogiLedSetLightingForKeyWithQuartzCode(int keyCode, int redPercentage, int greenPercentage, int bluePercentage);
+
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
+        public static extern bool LogiLedSetLightingForKeyWithKeyName(keyboardNames keyCode, int redPercentage, int greenPercentage, int bluePercentage);
+
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
+        public static extern bool LogiLedSetLightingForTargetZone(byte deviceType, int zone, int redPercentage, int greenPercentage, int bluePercentage);
+
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
+        public static extern bool LogiLedSaveLightingForKey(keyboardNames keyName);
+
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
+        public static extern bool LogiLedRestoreLightingForKey(keyboardNames keyName);
+
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
+        public static extern bool LogiLedFlashSingleKey(keyboardNames keyName, int redPercentage, int greenPercentage, int bluePercentage, int msDuration, int msInterval);
+
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
+        public static extern bool LogiLedPulseSingleKey(keyboardNames keyName, int startRedPercentage, int startGreenPercentage, int startBluePercentage, int finishRedPercentage, int finishGreenPercentage, int finishBluePercentage, int msDuration, bool isInfinite);
+
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
+        public static extern bool LogiLedStopEffectsOnKey(keyboardNames keyName);
+
+        [DllImport(dllpath, CallingConvention = CallingConvention.Cdecl)]
+        public static extern void LogiLedShutdown();
+    }
+}
