+++ Project-Aurora/Project-Aurora/Settings/Layers/WrapperLightsLayerHandler.cs
@@ -9,6 +9,7 @@ using Newtonsoft.Json;
 using Aurora.Devices;
 using Aurora.Settings.Overrides;
 using Aurora.Utils;
+using LedCSharp;
 
 namespace Aurora.Settings.Layers
 {
@@ -109,9 +110,9 @@ namespace Aurora.Settings.Layers
                 }
                 else
                 {
-                    Devices.Logitech.Logitech_keyboardBitmapKeys logi_key = Devices.Logitech.LogitechDevice.ToLogitechBitmap(key);
+                    var logi_key = Utils.DeviceKeysUtils.ToLogitechBitmap(key);
 
-                    if (logi_key != Devices.Logitech.Logitech_keyboardBitmapKeys.UNKNOWN && bitmap.Length > 0) {
+                    if (logi_key != Logitech_keyboardBitmapKeys.UNKNOWN && bitmap.Length > 0) {
                         var color = GetBoostedColor(Utils.ColorUtils.GetColorFromInt(bitmap[(int)logi_key / 4]));
                         bitmap_layer.Set(key, color);
 
@@ -238,16 +239,16 @@ namespace Aurora.Settings.Layers
             }
             else if (ngw_state.Command.Equals("SetLightingForKeyWithKeyName") || ngw_state.Command.Equals("SetLightingForKeyWithScanCode"))
             {
-                var bitmap_key = Devices.Logitech.LogitechDevice.ToLogitechBitmap((LedCSharp.keyboardNames)(ngw_state.Command_Data.key));
+                var bitmap_key = Utils.DeviceKeysUtils.ToLogitechBitmap((LedCSharp.keyboardNames)(ngw_state.Command_Data.key));
 
-                if (bitmap_key != Devices.Logitech.Logitech_keyboardBitmapKeys.UNKNOWN)
+                if (bitmap_key != Logitech_keyboardBitmapKeys.UNKNOWN)
                 {
                     bitmap[(int)bitmap_key / 4] = (int)(((int)ngw_state.Command_Data.red_start << 16) | ((int)ngw_state.Command_Data.green_start << 8) | ((int)ngw_state.Command_Data.blue_start));
                 }
             }
             else if (ngw_state.Command.Equals("FlashSingleKey"))
             {
-                Devices.DeviceKeys dev_key = Devices.Logitech.LogitechDevice.ToDeviceKey((LedCSharp.keyboardNames)(ngw_state.Command_Data.key));
+                Devices.DeviceKeys dev_key = Utils.DeviceKeysUtils.ToDeviceKey((LedCSharp.keyboardNames)(ngw_state.Command_Data.key));
                 LogiFlashSingleKey neweffect = new LogiFlashSingleKey(dev_key, Color.FromArgb(ngw_state.Command_Data.red_start, ngw_state.Command_Data.green_start, ngw_state.Command_Data.blue_start),
                         ngw_state.Command_Data.duration,
                         ngw_state.Command_Data.interval
@@ -261,7 +262,7 @@ namespace Aurora.Settings.Layers
             }
             else if (ngw_state.Command.Equals("PulseSingleKey"))
             {
-                Devices.DeviceKeys dev_key = Devices.Logitech.LogitechDevice.ToDeviceKey((LedCSharp.keyboardNames)(ngw_state.Command_Data.key));
+                Devices.DeviceKeys dev_key = Utils.DeviceKeysUtils.ToDeviceKey((LedCSharp.keyboardNames)(ngw_state.Command_Data.key));
                 long duration = ngw_state.Command_Data.interval == 0 ? 0 : ngw_state.Command_Data.duration;
 
                 LogiPulseSingleKey neweffect = new LogiPulseSingleKey(dev_key, Color.FromArgb(ngw_state.Command_Data.red_start, ngw_state.Command_Data.green_start, ngw_state.Command_Data.blue_start),
